:truck: rename SiteForms/ folder to /site-forms
Why? JS standard folder names are dash case to for cross compatibility with case insensitive file system such as Windows and Mac OS
