+++ src/web/src/components/Sites/site-forms/Photos.vue
